+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-17T02:34:22.765Z",
+    "2026-07-17T10:58:19.597Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {