+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -4042,7 +4042,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KXPYT7W4DGKMQVKD6N67FT3N"),
+  id: z.string().default("01KXQVMWAD9V5PM3TEAV06Y043"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),