+++ packages/orq-rc/RELEASES.md
@@ -10899,3 +10899,13 @@ Based on:
 - [typescript v4.12.0-rc.36] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.36] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.36 - packages/orq-rc
+
+## 2026-07-17 11:01:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.2 (2.918.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.37] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.37] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.37 - packages/orq-rc